Stop steps list regex matching newlines
The regex used `^\s*` to force the match to be anchored to the start of
a line followed by whitespace, but `\s` matches newlines as well as tabs
and spaces which causes the replaced markup to not be separated from the
content above by a double newline. See test for an example.

This change makes the regex use `[ \t]` to match acceptable whitespace
for the start of the line.